Fix kafka replication issues
Modified how the topics for Kafka are created to ensure that the replicas are respected.

This has the side-effect of hard coding it to 3 kafka replicas even further than it already is.  That may become a problem in the future, but for now seems like more of a nice to have.