BPS Refactor
Modified how the bps was being integrated to utilize disk to pass messages instead of relying on the http interface.  This was done because of performance issues with the network approach; namely, the majority of the work that BPS was doing was being ignored because there was only so much room in the queue it was generating.

Making this change breaks metrics pretty hard, but is functional other than that.